Диплом/Source - исходный код/college_system_100/college_system/requirements.txt
django==4.2.16
psycopg2-binary==2.9.10
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-allauth==0.63.3
pandas==2.2.3
openpyxl==3.1.5
reportlab==4.2.5
python-dotenv==1.0.1

